# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Gestion des Comptes",
        "description": "API pour la gestion des comptes bancaires",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Serveur de développement"
        },
        {
            "url": "/api/v1"
        }
    ],
    "paths": {
        "/comptes": {
            "get": {
                "tags": [
                    "Comptes"
                ],
                "summary": "Liste paginée des comptes",
                "description": "Récupère une liste paginée de tous les comptes avec possibilité de filtrage",
                "operationId": "5fdb1d9d6a1e1d81e65a694d33a249c6",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Numéro de la page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Nombre d'éléments par page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    },
                    {
                        "name": "type",
                        "in": "query",
                        "description": "Type de compte (cheque, epargne, courant)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "cheque",
                                "epargne",
                                "courant"
                            ]
                        }
                    },
                    {
                        "name": "statut",
                        "in": "query",
                        "description": "Statut du compte",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Recherche globale dans les comptes",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "Champ de tri",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "id"
                        }
                    },
                    {
                        "name": "order",
                        "in": "query",
                        "description": "Ordre de tri (asc ou desc)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "desc",
                            "enum": [
                                "asc",
                                "desc"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des comptes récupérée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "http_code": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Liste des comptes récupérée avec succès"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        },
                                        "pagination": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Comptes"
                ],
                "summary": "Crée un nouveau compte",
                "description": "Crée un nouveau compte bancaire",
                "operationId": "a896aedada2e34a114a6b2bda0468e82",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "numero_compte",
                                    "type",
                                    "solde_initial",
                                    "devise",
                                    "client_id"
                                ],
                                "properties": {
                                    "numero_compte": {
                                        "type": "string",
                                        "example": "CB001234567890"
                                    },
                                    "type": {
                                        "type": "string",
                                        "enum": [
                                            "cheque",
                                            "epargne",
                                            "courant"
                                        ],
                                        "example": "cheque"
                                    },
                                    "solde_initial": {
                                        "type": "number",
                                        "format": "decimal",
                                        "example": 1000
                                    },
                                    "devise": {
                                        "type": "string",
                                        "example": "FCFA"
                                    },
                                    "statut": {
                                        "type": "string",
                                        "example": "actif"
                                    },
                                    "client_id": {
                                        "type": "string",
                                        "example": "uuid-client"
                                    },
                                    "metadata": {
                                        "type": "object",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Compte créé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "http_code": {
                                            "type": "integer",
                                            "example": 201
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Compte créé avec succès"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Données invalides",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "http_code": {
                                            "type": "integer",
                                            "example": 400
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Erreur de validation"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}